[Add] /user/:id 페이지 기본 레이아웃 구현
